make index long

+++ 0x19-hash_tables/5-hash_table_print.c
@@ -7,9 +7,10 @@
  */
 void hash_table_print(const hash_table_t *ht)
 {
-	unsigned int i = 0;
+	unsigned long int i = 0;
 	hash_node_t *tmp;
 	unsigned int boo = 0;
+
 	if (ht == NULL)
 		return;
